[feat_master_scheduler][scheduler] fix scheduling bug

+++ taiga-scheduler/src/main/java/com/dtstack/taiga/scheduler/server/builder/JobSortWorker.java
@@ -12,5 +12,5 @@ public interface JobSortWorker {
      * 排序
      * @return 序号
      */
-    Long getSort();
+    Integer getSort();
 }
